+++ src/model/entry.ts
