+++ src/model/error/exception.ts
