+++ src/model/error/resolver-exception.ts
