+++ src/service/resolver/amazon/amazon-format-resolver.service.ts
@@ -2,9 +2,9 @@ import * as http from 'http';
 import * as https from 'https';
 import { HTMLElement } from 'node-html-parser';
 
-import { HtmlUtil } from '../html/html-util';
-import { NullableHtmlElement } from '../html/nullable-html-element';
-import { HttpUtil } from '../http/http-util';
+import { NullableHtmlElement } from '../../../model/html/nullable-html-element';
+import { HtmlUtil } from '../../../util/html-util';
+import { HttpUtil } from '../../../util/http-util';
 
 export class AmazonFormatResolverService {
   private static readonly KINDLE_UNLIMITED_ID = '.a-icon-kindle-unlimited';
