+++ src/service/resolver/audible/audible-information.ts
