+++ src/service/resolver/audible/audible-resolver.service.ts
@@ -1,11 +1,11 @@
 import { HTMLElement } from 'node-html-parser';
 import { URL } from 'url';
 
+import { NullableHtmlElement } from '../../../model/html/nullable-html-element';
+import { SiteResolver } from '../../../model/resolver/site-resolver.enum';
+import { Message } from '../../../model/telegram/message';
+import { HtmlUtil } from '../../../util/html-util';
 import { AbstractResolver } from '../abstract-resolver';
-import { NullableHtmlElement } from '../html/nullable-html-element';
-import { Message } from '../message';
-import { HtmlUtil } from './../html/html-util';
-import { SiteResolver } from './../site-resolver.enum';
 import { AudibleAuthor, AudibleInformation } from './audible-information';
 
 export class AudibleResolverService extends AbstractResolver {
